fix(nav): unify Preview badge styles
Share one PreviewBadge component between directory titles and TOC entries so their visual treatment stays identical.
