+++ src/components/Badge/PreviewBadge.tsx
@@ -1,0 +1,30 @@
+import Chip from "@mui/material/Chip";
+import { useTheme } from "@mui/material/styles";
+
+const PreviewBadge = (props: { label: string }) => {
+  const theme = useTheme();
+
+  return (
+    <Chip
+      label={props.label}
+      size="small"
+      variant="outlined"
+      sx={{
+        flexShrink: 0,
+        height: "20px",
+        fontSize: "12px",
+        fontWeight: 400,
+        borderRadius: "10px",
+        pointerEvents: "none",
+        "& .MuiChip-label": {
+          paddingLeft: "8px",
+          paddingRight: "8px",
+          lineHeight: "20px",
+          color: theme.palette.carbon[700],
+        },
+      }}
+    />
+  );
+};
+
+export default PreviewBadge;
